4. multi-modal-chatbot/backend/models/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
ollama==0.1.7
pillow==10.2.0
diffusers==0.27.2
transformers==4.39.1
torch==2.2.1
numpy==1.26.4
pydantic==2.6.3
python-dotenv==1.0.1
